requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
